backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy
pydantic==1.10.0a1
python-dotenv==1.0.0
alembic
Pillow==10.0.1
python-multipart==0.0.6
pycodestyle==2.11.0
mypy==1.5.1
pylint==2.17.5
psycopg2-binary